+++ src/core/shared/interfaces/webviewPlayerState.ts
@@ -1,0 +1,16 @@
+import * as alt from 'alt-shared';
+
+export interface WebViewPlayerState {
+    health: number;
+    armour: number;
+    weapon: number;
+    speed: number;
+    stamina: number;
+    pos: alt.IVector3;
+    rot: alt.IVector3;
+    micLevel: number;
+    isTalking: boolean;
+    isAiming: boolean;
+    inVehicle: boolean;
+    isDriver: boolean;
+}